// src/lib/apis/models/FinggerPlatform_Server_Shared_DataTransferObjects_Game_GameEventPublic.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { FinggerPlatform_Server_Shared_DataTransferObjects_Game_GameEventInformationPublic } from './FinggerPlatform_Server_Shared_DataTransferObjects_Game_GameEventInformationPublic';

export type FinggerPlatform_Server_Shared_DataTransferObjects_Game_GameEventPublic = {
  bundleId?: string | null;
  title?: string | null;
  imageUrl?: string | null;
  informations?: Array<FinggerPlatform_Server_Shared_DataTransferObjects_Game_GameEventInformationPublic> | null;
};