// src/lib/apis/models/FinggerPlatform_Server_Shared_DataTransferObjects_Notification_GameNotificationPublic.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type FinggerPlatform_Server_Shared_DataTransferObjects_Notification_GameNotificationPublic = {
  uuid?: string | null;
  readonly type?: string | null;
  title?: string | null;
  publishedAt?: string;
  hasBadge?: boolean;
  body?: string | null;
  imageUrl?: string | null;
};